mapping done

+++ src/main/java/com/example/demo/repository/UserRepository.java
@@ -14,8 +14,6 @@ public interface UserRepository extends JpaRepository<User, Integer> {
 	
 	public User findByUserId(Integer integer);
 
-	public User findOneByUserId(Long userId);
-	
-	//public boolean existUserByEmail(String email);
+	public User findOneByUserId(Integer userId);
 
 }
